Replace gyroscope with vibration sensor
The GY-521 (MPU6050) wasn't useful for detecting when the device was placed down and it was difficult to work with. A KY-002 vibration sesnor has been used instead. It also means there's no longer a need for a 10 second timer after movement was detected as we now know when the device was placed, not when it was picked up; the sensors are only meant to start reading when "movement stopped".