Update WiFi connectivity via BLE
The BLE WiFi Connection Service now has a characteristic that determines if the device is connected to WiFi, which will be useful for applications made to work with the device. The WiFi setup now only attempts to begin a connection when both an SSID and password are provided. The device will try to connect to the access point three times, with an interval of one second, before flagging the connection as failed.