Add setting temperature threshold via Bluetooth
BLE Characteristics to set the minimum and maximum healthy temperatures for a plant have been added. The callbacks for writing to a temperature Characterstic expect a string value. Because of issues with trying to use the C++ functions to parse the string to a float, it converts the string to a char array, then to an Arduino String. It uses the methods provided by the Arduino library to convert the String to a float value.

The Characterstic to determine if a device is connected to WiFi has been removed. This is going to be replaced with a server side flag. The BLEWrapper file has also been renamed to PlantagotchiBLE, it just makes life easier.