Update device connectivity and sensor reading
The major changes involve the splitting of the code base to the renamed Connectivity.cpp and Sensors.cpp to seperate points of concern, with both header files for these sources have being put into a shared namespace.

The functions to get data from the sensors have been combined to avoid code duplication. All variables relating to data read from sensors are now set to int32_t to account for the large values that can be read from the light sensor. This is to prevent potential overflows when calculating the average. Outliers in readings will now be removed from the data set. The average is now checked against the set range and not just the lower bound.

There's also been big changes to serial outputs to make debugging a better experience.